chore: update spec-update.spec-without-setup-funct
ion
