+++ tests/spec/spec-update.spec-without-setup-function.spec.ts
@@ -1,53 +1,37 @@
-import { Tree } from '@angular-devkit/schematics';
-import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
+import { LogEntry } from '@angular-devkit/core/src/logger';
 import { filter } from 'rxjs/operators';
-import { collectionPath } from './common';
+import { setupBase } from './common';
 
+const folder = 'spec-update-spec-without-setup-function';
+const file = 'spec-without-setup.component.ts';
 describe('Calling update on existing specs without setup function', () => {
-    let tree = Tree.empty();
-
-    beforeEach(() => {
-        tree = Tree.empty();
-        tree.create(
-            'c.ts',
-            `import { LogService, BDep } from '@angular/core';
-
-            export class C  {
-                constructor(
-                    private bDep: bDep,
-                    private logger: LogService
-                ) {}
-            `
-        );
-
-        tree.create(
-            'c.spec.ts',
-            `import { bDep } from '@angular/core';
-
-            describe('C', () => {
-            });`
-        );
-    });
-
-    it('should pass successfully', async  () => {
+    it('should pass successfully', async () => {
         // arrange
-        const runner = new SchematicTestRunner('schematics', collectionPath);
+        const { run, fullFileName, add, testFileName, letLogger } = setupBase(folder, file);
+        add(fullFileName);
+        add(testFileName);
+
         // act
-        const errors = [];
-        runner.logger.pipe(filter(v => v.level === 'error')).subscribe(v => errors.push(v));
-        await runner.runSchematicAsync('spec', { name: './c.ts', update: true }, tree).toPromise();
+        const errors: LogEntry[] = [];
+        letLogger((logger) =>
+            logger.pipe(filter((v) => v.level === 'error')).subscribe((v) => errors.push(v))
+        );
+        await run({ name: fullFileName, update: true }).catch((e) => {});
         // assert
         expect(errors.length).toBe(0);
     });
 
-    it('should create the setup function and then update it', async  () => {
+    it('should create the setup function and then update it', async () => {
         // arrange
-        const runner = new SchematicTestRunner('schematics', collectionPath);
+        const { run, fullFileName, add, testFileName, } = setupBase(folder, file);
+        add(fullFileName);
+        add(testFileName);
+
         // act
-        const result = await runner.runSchematicAsync('spec', { name: './c.ts', update: true }, tree).toPromise();
+        const result = await run({ name: fullFileName, update: true });
         // assert
         // @ts-ignore
-        const contents = result.readContent('./c.spec.ts');
+        const contents = result.readContent(testFileName);
         // update should add LogService to imports, to construct params and create a spy for it
         expect(contents).toContain(`function setup() {`);
         expect(contents).toContain("import { bDep, LogService } from '@angular/core';");
