+++ tests/spec/test-data/spec-update-spec-without-setup-function/spec-without-setup.component.spec.ts
@@ -1,0 +1,4 @@
+import { bDep } from '@angular/core';
+
+describe('C', () => {
+});
