+++ tests/spec/test-data/spec-update-spec-without-setup-function/spec-without-setup.component.ts
@@ -1,0 +1,7 @@
+import { LogService, BDep } from '@angular/core';
+
+export class C  {
+    constructor(
+        private bDep: bDep,
+        private logger: LogService
+    ) {}
